refactor(ui): source LATENCY_VALIDATED_ON_CONNECT from generated, drop hand-mirror
